Removing "cry for help"
I figured out that VSCode breakpoints don't work in the lambda'd functions (or maybe Beam pipelines.  The code-writing-code that is good for performance is very hard to debug in).  I see there python is writing temporary .py files.  So there your "breakpoint()" was my savior, because that got replicated inside the temporay .py files. Which the VScode breakpoints did not.